prior volume random variable notes
Added notes on how importance sampling works on a statistical level. This includes a discussion of order statistics and involves explicit results for the distribution of the prior volumes during an ideal run. Importantly, this also includes the joint distribution of the remaining K live points at the end of a run with N iterations.

Changes to pnestle.py just reverting the previous commit since there was a serious bug in the implementation. So now the parallelization is just the simple 'queue up a bunch of points and dump any extra when you need to update', which should be fine I think.