error accuracy checks
Using the order statistics derivations from earlier, I implemented a function that samples from the log(prior volume) distributions and propogates those results to compute the volume derivative estimates, log(weights), and log(evidence). I checked results on a 10-D pair of Gaussian shells, which looked good.

Results confirm that in most applications the first/second geometric moment approximations should more than suffice, and that errors in the importance weights appear to be effectively convolved with the posterior over the course of a single run.

One worthwhile note is that there is a simple treatment to include the final set of live points using a substantially better approximation than is currently included in nestle. Maybe worth implementing for completeness.